Title: 학점 계산기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231230/학점 계산기/credit-calculator.c
@@ -10,12 +10,12 @@ int main() {
         scanf("%f",&a[i]);
         sum+=a[i];
     }
-    printf("%.1f\n",sum/n);
+    printf("%.1lf\n",sum/n);
     if((float)sum/n>=4){
         printf("Perfect");
 
     }
-    else if((float )sum/n>=3) printf("Good");
+    else if(sum/n>=3) printf("Good");
     else printf("Poor");
     return 0;
 }
